Fix bugs with generators.

+++ Randomizer.OutputTests/Tests/Integer/IntegerInRangeOutputTest.cs
@@ -20,7 +20,7 @@ namespace Randomizer.OutputTests.Tests.Integer
 
             for (int i = 0; i < ExecutionTimes; i++)
             {
-                float randomValue = randomLong.GenerateValue(minValue, maxValue);
+                int randomValue = randomLong.GenerateValue(minValue, maxValue);
                 if (randomValue > maxValue || randomValue < minValue)
                 {
                     wrongResults.Add(randomValue.ToString(CultureInfo.InvariantCulture));
