+++ Randomizer.OutputTests/Tests/Long/LongPositiveValueOutputTest.cs
@@ -15,7 +15,7 @@ namespace Randomizer.OutputTests.Tests.Long
 
             for (int i = 0; i < ExecutionTimes; i++)
             {
-                float randomValue = randomLong.GeneratePositiveValue();
+                long randomValue = randomLong.GeneratePositiveValue();
                 if (randomValue < 0)
                 {
                     wrongResults.Add(randomValue.ToString(CultureInfo.InvariantCulture));
